+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Includes potentially breaking changes.
 
 - Renamed tabs and runtime event handlers for consistency, added deprecated wrapper for the old function names.  
 - storage/get will now keywordize the keys on the dictionary being returned.
+- Added extra parameters for storage functions to be able to indicate the area.
+- BREAKING CHANGE: Default area for storage functions will be `local`.
 
 
 ## 0.0.4
@@ -14,7 +16,7 @@ Hotfix for a khroma/util.cljs bug.
 
 ## 0.0.3
 
-- New storage namespace. Wraps storage functions and events, currently all invocations go to sync (need to decide if we will provide separate functions for local/sync).
+- New storage namespace. Wraps storage functions and events.
 - New browser namespace. Currently has only an on-clicked event for receiving click notifications on the extension button. Will be used to wrap browserAction functions.
 - New windows namespace. Will be used to wrap windows functions.
 - Miscellaneous refactoring and documentation improvements.